pyproject.toml
[project]
name = "stars-site-back"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "cryptography>=45.0.6",
    "django>=5.2.5",
    "django-stubs>=5.2.2",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "mysqlclient>=2.2.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytelegrambotapi>=4.28.0",
    "python-dotenv>=1.1.1",
    "pythoniq>=0.0.1",
    "redis>=6.4.0",
    "tonutils>=0.5.2",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.10",
]
